.conductor/albany/requirements.txt
# Requirements for RAG Evaluation Framework

ragas
langchain
langchain_openai
langchain_community
python-dotenv
pandas
numpy
openpyxl  # For Excel file processing
python-docx  # For Word document processing
PyPDF2  # For PDF processing
pdfplumber  # Alternative PDF processing library
pillow  # For image processing
pytesseract  # For OCR
streamlit  # For manual evaluation interface
matplotlib  # For result visualization
seaborn  # For statistical data visualization